app/notifications_service/requirements.txt
Django
channels
gunicorn
psycopg2-binary
mysqlclient
channels_redis
django-environ
djangorestframework
whitenoise
psycopg2
daphne
requests